Implemented baseline model
Implemented the binning method to capture different hyper-rectangles that the curves pass through  as a binary vector. For the different binary vectors per speaker a randomized subclass method is applied - the technique uses permutations of the  order of vectors belonging to a positive class and subsequently applies AND operations to construct exclusive and maximum sets which create a "signature" for a given speaker. Unknown samples are then classified by matching the "signature" or being close to it. A lot of simple tests are provided, as well as comments and visualizations of the final rectangle rules.